requirements.txt
streamlit
gspread
google-auth
plotly
pandas
certifi